requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
deprecation==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.7.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.37.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
python-keycloak==3.7.0
opencv-python==4.9.0.80
dlib==19.24.2
pytesseract==0.3.10
matplotlib==3.8.3
scikit-image==0.22.0
easyocr==1.7.1
google-generativeai==0.5.2
pillow==10.2.0